medium-mcp/pyproject.toml
[project]
name = "medium-mcp"
version = "0.1.0"
description = "MCP server for reading Medium premium articles via Playwright"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "playwright>=1.44.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.scripts]
medium-mcp = "medium_mcp.server:main"
medium-mcp-setup-auth = "medium_mcp.setup_auth:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medium_mcp"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
